+++ lib/ember/filters/haml.rb
