+++ lib/ember/filters/slim.rb
