+++ lib/ember/handlebars/template.rb
@@ -1,3 +1,5 @@
+require "ember/handlebars/source"
+
 module Ember
   module Handlebars
     class Template < Tilt::Template
@@ -10,7 +12,7 @@ module Ember
       def evaluate(scope, locals, &block)
         template = mustache_to_handlebars(scope, data)
 
-        if ::Rails.configuration.handlebars.precompile
+        if configuration.precompile
           func = Ember::Handlebars.compile(template)
           "Ember.TEMPLATES[#{template_path(scope.logical_path).inspect}] = Ember.Handlebars.template(#{func});\n"
         else
@@ -30,12 +32,15 @@ module Ember
 
       def template_path(path)
         path = path.split('/')
+        root = configuration.template_root
 
-        unless ::Rails.configuration.handlebars.template_root.blank?
-          path.delete(::Rails.configuration.handlebars.template_root)
-        end
+        path.delete(root) unless root.blank?
 
-        path.join(::Rails.configuration.handlebars.template_path_separator)
+        path.join(configuration.template_path_separator)
+      end
+
+      def configuration
+        ::Rails.configuration.handlebars
       end
 
       def indent(string)
