+++ lib/ember/handlebars/version.rb
@@ -1,0 +1,5 @@
+module Ember
+  module Handlebars
+    VERSION = '1.0.beta.2'
+  end
+end
