+++ lib/ember_rails.rb
@@ -1,25 +1,7 @@
 require 'sprockets'
 require 'sprockets/engines'
 
-require "ember/handlebars/source"
-require "ember/handlebars/template"
+require "ember/rails/engine"
 
-require "ember_rails/slim" if defined? Slim
-require "ember_rails/haml" if defined? Haml
-
-module Ember
-  module Rails
-    class Engine < ::Rails::Engine
-      config.handlebars = ActiveSupport::OrderedOptions.new
-      config.handlebars.precompile = ::Rails.env.production?
-      config.handlebars.template_root = nil
-      config.handlebars.template_path_separator = '/'
-
-      initializer :setup_ember_rails, :group => :all do |app|
-        app.assets.register_engine '.handlebars', Ember::Handlebars::Template
-        app.assets.register_engine '.hbs', Ember::Handlebars::Template
-        app.assets.register_engine '.hjs', Ember::Handlebars::Template
-      end
-    end
-  end
-end
+require "ember/filters/slim" if defined? Slim
+require "ember/filters/haml" if defined? Haml
